Karazhan 2: Electric Boogaloo
Apocalypse will now either spawn if Laith or Ariden are alive.
Added holy sites for Throssians and Shadowlanders.
Moroes is now playable in 603 bookmark as an independent duke-tier ruler of Deadwind Pass.
Ariden is now playable in 603 bookmark as a count-tier under Moroes.
The provinces of Deadwind Pass now have Shadowlander religion in 603.