Blasted Lands and the Stormreaver Remnant
Teron'gor is now properly missing in 603, as well as has proper traits and gfxculture.
Fel'dan is now a fel orc in 603 and is a courtier under Jubei'thos.
Fel'dan's succubi are now implemented.
Added Sevine, Grol and Alistarj.
The Blasted Lands underwent a rework in 603, being divided between the still feuding Shadowsworn Legion-aligned warlords, with Nethergarde barely holding to the Dark Portal and the Nethergarde Keep.
Kazzak has been commented out as he's going to serve a different purpose during the Legion Invasion from now on.