Worgen event chain reworked
Worgen event chain can now fire. The Book of Ur is used by mages, and the Scythe of Elune is used by Druids. The chain only requires that a courtier or vassal has the artifact, rather than the liege, so that it has a reasonable chance of firing. Neighboring the scourge greatly reduces the number of days to fire.

-Scythe of Elune spawn added to Howling Vale
-Arugal character entry tidied up