Closed #1185 cooldowns
Explained to the AI when to do genocides and made AI's chances of clicking them non-random. The AI doesn't genocide if there's 4 or more ongoing genocides in the AI's realm. It should make genocides more common and less dangerous to AI's realms.
The AI no longer genocides provinces during war.