"fixed" history for cavefall,
I had replaced Thug and Pippin (who were originally owners of this county before war of three hammers started. Once that start date is being worked on this will need to be fixed more) with Garfang and Gavin. So then in the history of the province it showed they owned it.