silverpine, deathknell
filled up silverpine, no owner of d_deep_elm yet so they are all count vassals of k_lordaeron
deathknell is home to the last amani trolls in lordaeron. they are in c_deathknell and deathknell is called Agor'votha, and as a result of being right above Zakajz the Corruptor, the ruling dynasty secretly worships Yogg-Saron. These trolls are conquered by lordaeron in 536.5.21